requirements.txt
django
Pillow==4.0.0
django-braces==1.10.0
django-extensions==1.7.5
restless
djangorestframework
django-rest-swagger
django-filter
django-oauth-toolkit
django-rest-framework-social-oauth2
python-social-auth
social-auth-app-django
django-cors-headers